+++ python/openmldb/test/case_conf.py
@@ -19,5 +19,5 @@
 OpenMLDB address used in test
 """
 
-OpenMLDB_ZK_CLUSTER="127.0.0.1:6181"
-OpenMLDB_ZK_PATH="/onebox"
+OpenMLDB_ZK_CLUSTER="127.0.0.1:2181"
+OpenMLDB_ZK_PATH="/openmldb"
