+++ python/openmldb/test/sdk_smoke_test.py
